[FE] refactor: 받은 요청,  보낸 요청, 받은 요청에 수락/거절 로직 훅으로 분리 (#89)
