+++ src/frontend/src/components/friend/PendingFriendsList/hooks/useReceivedRequests.ts
@@ -1,0 +1,14 @@
+import { useQuery } from '@tanstack/react-query';
+
+import { getReceivedRequest } from '@/api/friends';
+
+const useReceivedRequests = () => {
+  const result = useQuery({ queryKey: ['receivedRequests'], queryFn: getReceivedRequest });
+
+  return {
+    receivedRequests: result.data,
+    ...result,
+  };
+};
+
+export default useReceivedRequests;
